test: clean up and txn test
