+++ src/common/meta/src/kv_backend/memory.rs
@@ -325,7 +325,9 @@ mod tests {
     use crate::error::Error;
     use crate::kv_backend::test::{
         prepare_kv, test_kv_batch_delete, test_kv_batch_get, test_kv_compare_and_put,
-        test_kv_delete_range, test_kv_put, test_kv_range, test_kv_range_2,
+        test_kv_delete_range, test_kv_put, test_kv_range, test_kv_range_2, test_txn_compare_equal,
+        test_txn_compare_greater, test_txn_compare_less, test_txn_compare_not_equal,
+        test_txn_one_compare_op, text_txn_multi_compare_op,
     };
 
     async fn mock_mem_store_with_data() -> MemoryKvBackend<Error> {
@@ -382,5 +384,16 @@ mod tests {
         let kv_backend = mock_mem_store_with_data().await;
 
         test_kv_batch_delete(kv_backend).await;
+    }
+
+    #[tokio::test]
+    async fn test_memory_txn() {
+        let kv_backend = Arc::new(MemoryKvBackend::<Error>::new());
+        test_txn_one_compare_op(kv_backend.clone()).await;
+        text_txn_multi_compare_op(kv_backend.clone()).await;
+        test_txn_compare_equal(kv_backend.clone()).await;
+        test_txn_compare_greater(kv_backend.clone()).await;
+        test_txn_compare_less(kv_backend.clone()).await;
+        test_txn_compare_not_equal(kv_backend).await;
     }
 }
